Move more init out of bootstrap
